Crate medici_core

Types for implementing a state machine for (board) games. This crate provides an opinionated framework which the developers can use for their own games.

Re-exports

pub extern crate value_from_type_traits;

pub extern crate failure;

Modules

ctstack
Module which implements a CONS-LIST for transition validation at compile time.

error
Types, to be used within the system, providing context of unexpected behaviour.

function
Contains the core functionality items for our system.

macros
Defines useful macro's to simplify syntax of implementations

marker
Primitive traits which can be used as constraints by the core components.

prefab
Types with ubiquitous usage within state machines (containers) or specific to the opiniated nature of this framework. Framework users are encouraged to use these types whenever possible. The services are built to make use of these types, but freedom was preserved wherever possible.

service
Types that provide functionality to your application. These types can be added AD-HOC to your (state) machine declaration. Other components of this framework also need some of these services to be implemented before use.

stm
Traits enforcing state machine behaviour.

storage
Types with each different functionality to store data.

transaction
Core functionality helper methods.

Macros

build_exec_triggers_checked
Macro used for building a function that automatically constructs a method called [exec_triggers].

ct
Usability macro for pushing a new type onto the CTStack.

ctxt

hydrate