Module medici_core::function

Contains the core functionality items for our system.

Structs

CardId
Type that's generally used to identify and order Card objects.

Traits

ArrayStorageCompliance
Defines array access behaviour for storage objects.

Card
Trait representing an actual game card.

CardBuilder
Trait used to create a new Card object.

EffectState
Trait generalizing over any state which is used to bootstrap an execution of triggers.

Entity
Trait representing an object which properties can be altered dynamically (at runtime).

EntityBuilder
Trait used to create a new Entity object.

Identifiable
Trait enforcing implementing objects to expose their identifier. The uniqueness of this identifier depends on the implementing object itself.

IndexedStorageCompliance
Defines indexed behaviour for a certain storage object.

ServiceCompliance
Trait for implementing a certain service on the state machine.

StackStorageCompliance
Defines stack behaviour for a certain storage object.

State
Trait generalizing over any state that's present in the state machine.

StateContainer
Trait generalizing over any structure that could act as a container of states.

TriggerState
Trait generalizing over any state that's used to pass into trigger callbacks when trigger conditions are met.

Type Definitions

EntityId
Type that's generally used to identify and order Entity objects.