Trait medici_core::function::Card

pub trait Card: Identifiable {
    type TimingEnum: TimingEnumerator;
    type TriggerEnum: TriggerEnumerator;
}

Trait representing an actual game card.

A card is an Entity but it's usage is semantically disjunct enough to warrant a seperate type.

Identifiable

Cards are identified with GLOBAL UNIQUE identifiers. This allows for cards to be constructed once and a static table could be generated to retrieve a reference from there. It's also possible to hand out owned cards given their identifier doesn't clash with any other card.

Note

This trait MUST ALWAYS be object safe! This provides the flexibility to store a bunch of Cards into one container.

Associated Types

type TimingEnum: TimingEnumerator

All timing types this card holds listeners for.

type TriggerEnum: TriggerEnumerator

All trigger types this card holds listeners for.