Trait medici_core::function::StateContainer

pub trait StateContainer {
    type State: State;
    type Transaction: Transaction = <Self::State as State>::Transaction;
    type TimingEnum: TimingEnumerator;
    type TriggerEnum: TriggerEnumerator;
    type TransitionRecord: CTStack;
}

Trait generalizing over any structure that could act as a container of states.

This container of states could be reworded as 'the state machine' itself.

Associated Types

type State: State

Type of the current state held by the state machine.

type Transaction: Transaction = <Self::State as State>::Transaction

Type of transaction object necessary to transition into the current state of the machine.

type TimingEnum: TimingEnumerator

Type which enumerates all possible timings contained by the machine.

type TriggerEnum: TriggerEnumerator

Type which enumerates all possible triggers contained by the machine.

type TransitionRecord: CTStack

Type representing the stack of types where the container state was transitioned in a pushdown manner.