Module medici_core::prefab

Types with ubiquitous usage within state machines (containers) or specific to the opiniated nature of this framework. Framework users are encouraged to use these types whenever possible. The services are built to make use of these types, but freedom was preserved wherever possible.

Note

The child (and grandchild) types defined within this module are built to be as generic as possible, but some make use of other types within the [prefab] module. Incompatibility may arise when the coupling with these other types is not considered.

Modules

card
Module containing items to work with game cards.

entity
Module containing standard entity structures.

prototype
Module containing standard prototypes.

runtime
Module containing methods which make working with the state machine a bit easier.

state
Module containing often used state types.

timing
Module containing ready-to-use types which can be used to construct trigger constraints.

transaction
Types for provisioning the next state when performing a transition within the state machine.

trigger
Module containing ready-to-use types which can be used to construct trigger constraints.