Struct medici_core::prefab::card::CardStruct

pub struct CardStruct<S, ETM, ETR> where
    S: Clone + Eq + Hash,
    ETM: TimingEnumerator + PartialEq + Copy,
    ETR: TriggerEnumerator + PartialEq + Copy,  {
    pub name: &'static str,
    pub state: HashMap<S, u32>,
    pub triggers: TriggerService<ETM, ETR>,
    // some fields omitted
}

Type representing a game card.

Fields
name: &'static str

Holds the name of this card.

state: HashMap<S, u32>

Holds the properties attributed to this card.

triggers: TriggerService<ETM, ETR>

Holds all triggers registered on this card.

Methods

impl<S, ETM, ETR> CardStruct<S, ETM, ETR> where
S: Debug + Clone + Eq + Hash,
ETM: TimingEnumerator + PartialEq + Copy,
ETR: TriggerEnumerator + PartialEq + Copy,

pub fn get_value_default(&self, key: &S) -> u32

Retrieves the value of the requested property defined within this card. 0 is returned as default value when the property key was not found!

pub fn get_value(&self, key: &S) -> Result<u32, MissingPropertyError<CardId, S>>

Retrieves the value of the requested property defined within this card.

pub fn set_value(&mut self, key: S, value: u32) -> Option<u32>

Store the provided property key with corresponding value into this card.

The old value is returned if the key was already known within this card.