Struct medici_core::prefab::entity::EntityStruct

pub struct EntityStruct<S, P> where
    S: Clone + Eq + Hash,
    P: ProtoEnumerator + Clone + Eq + Hash,  {
    pub state: HashMap<S, u32>,
    pub prototypes: HashSet<P>,
    pub human_readable: Option<String>,
    // some fields omitted
}

Type representing a stateful 'thing' within the state-machine.

Essentially everything can be classified as an entity because there are always properties that can be attributed to these 'things'. Next to properties, which can be freely manipulated, behaviours can also be defined. Behaviours are essentially recipes of property manipulations. These behaviours are called [Prototypes].

Fields
state: HashMap<S, u32>

Contains all properties attributed to this entity.

prototypes: HashSet<P>

Contains a set of behaviours that are attributed to this entity.

human_readable: Option<String>

Provides the option to assign a string to this entity.

Example usages include; Name of player entities, Short identity for debugging purposes.

Methods

impl<S, P> EntityStruct<S, P> where
S: Debug + Clone + Eq + Hash,
P: ProtoEnumerator + Debug + Eq + Hash + Clone,

pub fn get_value_default(&self, key: &S) -> u32

Retrieves the value of the requested property defined within this entity. 0 is returned as default value when the property key was not found!

pub fn get_value(
&self,
key: &S
) -> Result<u32, MissingPropertyError<EntityId, S>>

Retrieves the value of the requested property defined within this entity.

pub fn set_value(&mut self, key: S, value: u32) -> Option<u32>

Store the provided property key with corresponding value into this entity.

The old value is returned if the key was already known within this entity.

pub fn add_proto<PT>(&mut self) where
PT: Prototype + IntoEnum<P>,

Attach new behaviour to this specific entity.

pub fn remove_proto<PT>(&mut self) where
PT: Prototype + IntoEnum<P>,

Removes behaviour from this specific entity.

pub fn as_proto<'a, PT>(
&'a self
) -> Result<PT, MissingPrototypeError<EntityId, P>> where
PT: Prototype + IntoEnum<P> + From<&'a Self>,

Return this entity as the requested prototype.

pub fn as_proto_mut<'a, PT>(
&'a mut self
) -> Result<PT, MissingPrototypeError<EntityId, P>> where
PT: Prototype + IntoEnum<P> + From<&'a mut Self>,

Return this entity as the requested prototype.