Function medici_core::prefab::runtime::exec_trigger_stepped

pub unsafe fn exec_trigger_stepped<'a, 'b, M, TM, TR, ETM, ETR, I>(
    machine: M, 
    triggers: I
) -> Result<M, MachineError> where
    M: StateContainer + Borrow<M> + Debug + Clone + Send + Sync + 'static,
    M::State: TriggerState<Timing = TM, Trigger = TR>,
    TM: Timing + IntoEnum<ETM>,
    TR: Triggerable + IntoEnum<ETR>,
    ETM: TimingEnumerator + PartialEq + Copy,
    ETR: TriggerEnumerator + PartialEq + Copy,
    I: IntoIterator<Item = UnsafeTrigger<ETM, ETR>>, 

Executes all passed down triggers for the provided machine.

Unsafe

Verify that each provided UnsafeTrigger has been specifically created for the provided machine!