Struct medici_core::service::trigger::TriggerWrapper

pub struct TriggerWrapper<M, ETM, ETR> where
    M: StateContainer,
    M::State: TriggerState,
    <M::State as TriggerState>::Timing: Timing + IntoEnum<ETM>,
    <M::State as TriggerState>::Trigger: Triggerable + IntoEnum<ETR>,
    ETM: TimingEnumerator + PartialEq + Copy,
    ETR: TriggerEnumerator + PartialEq + Copy,
    fn(_: M) -> Result<M, MachineError>: Send + Sync,  { /* fields omitted */ }

Safe abstraction over UnsafeTrigger objects.

Methods

impl<M, ETM, ETR> TriggerWrapper<M, ETM, ETR> where
M: StateContainer,
M::State: TriggerState,
<M::State as TriggerState>::Timing: Timing + IntoEnum<ETM>,
<M::State as TriggerState>::Trigger: Triggerable + IntoEnum<ETR>,
ETM: TimingEnumerator + PartialEq + Copy,
ETR: TriggerEnumerator + PartialEq + Copy,

pub fn new(cb: fn(_: M) -> Result<M, MachineError>) -> Self

Constructs a new trigger method wrapper from the provided method.

pub unsafe fn try_from_trigger_entry(
x: UnsafeTrigger<ETM, ETR>
) -> Result<Self, TriggerFail>

Build a safe wrapper from a UnsafeTrigger object.

Safety

The exact StateContainer (state machine) is removed from the UnsafeTrigger. A safe wrapper can thus be generated for ANY state machine which TriggerState-associated types Timing and Trigger match on the requested machine.

See TriggerService for more information!

pub fn into_callback(self) -> fn(_: M) -> Result<M, MachineError>

Consumes this wrapper to retrieve the callback it contains.