Module medici_core::stm::checked

(Type) Checked transitions for state machines.

Traits

PullupFrom
Trait defining the contract for two-way transitions of the state machine. The Transaction object of the outgoing state is stored for re-use. This kind of transition will write state change history into the compile-time stack (CTStack).

PullupInto
Syntax simplifying trait in accordance to PullupFrom.

PushdownFrom
Trait defining the contract for two-way transitions of the state machine. The transition this trait defines is meant as a temporary one where the control flow will result in the state machine eventually performs a PullupFrom transition. Effectively reversing the Pushdown. The Transaction object of the outgoing state is stored for re-use. This kind of transition will write state change history into the compile-time stack (CTStack).

PushdownInto
Syntax simplifying trait in accordance to PushdownFrom.

TransitionFrom
Trait defining the contract for one-way transitions of the state machine. This kind of transition preserves the state change history.

TransitionInto
Syntax simplifying trait in accordance to TransitionFrom.